fix(test): move scope-review-modal tests and address code review issues
- Move test file from views/conn/__tests__/ to views/zero-page/__tests__/
  to match location of the source components being tested
- Remove duplicate ConnectorPermissionDialog tests already covered by
  connector-permission-dialog.test.tsx
- Replace CSS class assertion (.shrink-0) with semantic img query (CONN-D-033)
- Replace pure text label assertions with behavioral assertions:
  CONN-D-036: verify Reconnect and Close buttons are present
  CONN-S-037: verify Reconnect button is absent in loading state
  CONN-C-038: verify Reconnect button is absent in error state
- Fix non-null assertion in CONN-I-040 using safe guard clause
- Remove unnecessary as ConnectorType casts
