+++ turbo/apps/platform/src/views/zero-page/__tests__/connector-permission-scope.test.tsx
@@ -1,0 +1,236 @@
+/**
+ * Tests for ScopeReviewModal component.
+ *
+ * Tests page-level behavior via setupPage following platform testing principles:
+ * - Entry point: setupPage({ path: "/connectors" })
+ * - Mock (external): Web API via MSW
+ * - Real (internal): All signals, components, rendering
+ */
+
+import { afterEach, describe, expect, it, vi } from "vitest";
+import { screen, waitFor, within } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
+import { http, HttpResponse } from "msw";
+import { server } from "../../../mocks/server.ts";
+import { testContext } from "../../../signals/__tests__/test-helpers.ts";
+import { setupPage } from "../../../__tests__/page-helper.ts";
+import { setScopeReviewType$ } from "../../../signals/zero-page/settings/connectors.ts";
+import type { ConnectorType } from "@vm0/core";
+
+const context = testContext();
+
+afterEach(() => {
+  vi.restoreAllMocks();
+});
+
+async function openScopeReviewModal(
+  connectorType: ConnectorType,
+  scopeDiff: {
+    addedScopes: string[];
+    removedScopes: string[];
+    currentScopes: string[];
+    storedScopes: string[];
+  },
+) {
+  server.use(
+    http.get("*/api/zero/connectors/:type/scope-diff", () => {
+      return HttpResponse.json(scopeDiff);
+    }),
+  );
+  await setupPage({ context, path: "/connectors" });
+  await waitFor(() => {
+    expect(
+      screen.getByRole("heading", { name: "Connectors" }),
+    ).toBeInTheDocument();
+  });
+  context.store.set(setScopeReviewType$, connectorType);
+}
+
+describe("scope review modal - display", () => {
+  it("shows connector icon and label (CONN-D-033)", async () => {
+    await openScopeReviewModal("github", {
+      addedScopes: ["repo"],
+      removedScopes: [],
+      currentScopes: [],
+      storedScopes: [],
+    });
+
+    await waitFor(() => {
+      expect(screen.getByRole("dialog")).toBeInTheDocument();
+    });
+
+    // Label and ConnectorIcon img are rendered inside the dialog
+    const dialog = screen.getByRole("dialog");
+    expect(dialog).toHaveTextContent("GitHub");
+    expect(dialog.querySelector("img")).toBeInTheDocument();
+  });
+
+  it("shows added scopes with + prefix (CONN-D-034)", async () => {
+    await openScopeReviewModal("github", {
+      addedScopes: ["repo", "project"],
+      removedScopes: [],
+      currentScopes: [],
+      storedScopes: [],
+    });
+
+    await waitFor(() => {
+      const plusItems = screen.getAllByText("+");
+      expect(plusItems.length).toBeGreaterThan(0);
+    });
+    expect(screen.getByText("repo")).toBeInTheDocument();
+    expect(screen.getByText("project")).toBeInTheDocument();
+  });
+
+  it("shows removed scopes with - prefix (CONN-D-035)", async () => {
+    await openScopeReviewModal("github", {
+      addedScopes: [],
+      removedScopes: ["read:user"],
+      currentScopes: [],
+      storedScopes: [],
+    });
+
+    await waitFor(() => {
+      const minusItems = screen.getAllByText("-");
+      expect(minusItems.length).toBeGreaterThan(0);
+    });
+    expect(screen.getByText("read:user")).toBeInTheDocument();
+  });
+
+  it("shows Reconnect and Close buttons when scope data is loaded (CONN-D-036)", async () => {
+    await openScopeReviewModal("github", {
+      addedScopes: ["repo"],
+      removedScopes: [],
+      currentScopes: [],
+      storedScopes: [],
+    });
+
+    await waitFor(() => {
+      expect(
+        screen.getByRole("button", { name: "Reconnect" }),
+      ).toBeInTheDocument();
+    });
+    const dialog = screen.getByRole("dialog");
+    const closeButtons = within(dialog).getAllByRole("button", {
+      name: "Close",
+    });
+    const textCloseButton = closeButtons.find((btn) => {
+      return btn.textContent?.trim() === "Close";
+    });
+    expect(textCloseButton).toBeInTheDocument();
+  });
+});
+
+describe("scope review modal - states", () => {
+  it("loading state shows dialog without Reconnect button (CONN-S-037)", async () => {
+    server.use(
+      http.get("*/api/zero/connectors/:type/scope-diff", () => {
+        return new Promise<never>(() => {
+          // Never resolves — keeps component in loading state
+        });
+      }),
+    );
+
+    await setupPage({ context, path: "/connectors" });
+    await waitFor(() => {
+      expect(
+        screen.getByRole("heading", { name: "Connectors" }),
+      ).toBeInTheDocument();
+    });
+    context.store.set(setScopeReviewType$, "github");
+
+    await waitFor(() => {
+      expect(screen.getByRole("dialog")).toBeInTheDocument();
+    });
+    expect(
+      screen.queryByRole("button", { name: "Reconnect" }),
+    ).not.toBeInTheDocument();
+  });
+
+  it("error state shows dialog without Reconnect button (CONN-C-038)", async () => {
+    server.use(
+      http.get("*/api/zero/connectors/:type/scope-diff", () => {
+        return HttpResponse.json(
+          { message: "Internal error" },
+          { status: 500 },
+        );
+      }),
+    );
+
+    await setupPage({ context, path: "/connectors" });
+    await waitFor(() => {
+      expect(
+        screen.getByRole("heading", { name: "Connectors" }),
+      ).toBeInTheDocument();
+    });
+    context.store.set(setScopeReviewType$, "github");
+
+    await waitFor(() => {
+      expect(screen.getByRole("dialog")).toBeInTheDocument();
+    });
+    expect(
+      screen.queryByRole("button", { name: "Reconnect" }),
+    ).not.toBeInTheDocument();
+  });
+});
+
+describe("scope review modal - interactions", () => {
+  it("reconnect button triggers connector reconnection (CONN-I-039)", async () => {
+    const openSpy = vi
+      .spyOn(window, "open")
+      .mockReturnValue({ closed: true } as Window);
+
+    await openScopeReviewModal("github", {
+      addedScopes: ["repo"],
+      removedScopes: [],
+      currentScopes: [],
+      storedScopes: [],
+    });
+
+    await waitFor(() => {
+      expect(
+        screen.getByRole("button", { name: "Reconnect" }),
+      ).toBeInTheDocument();
+    });
+
+    const user = userEvent.setup();
+    await user.click(screen.getByRole("button", { name: "Reconnect" }));
+
+    await waitFor(() => {
+      expect(openSpy).toHaveBeenCalledWith(
+        expect.stringContaining("/api/zero/connectors/github/authorize"),
+        "_blank",
+        expect.any(String),
+      );
+    });
+  });
+
+  it("close button closes the ScopeReviewModal (CONN-I-040)", async () => {
+    await openScopeReviewModal("github", {
+      addedScopes: ["repo"],
+      removedScopes: [],
+      currentScopes: [],
+      storedScopes: [],
+    });
+
+    await waitFor(() => {
+      expect(screen.getByRole("dialog")).toBeInTheDocument();
+    });
+
+    const user = userEvent.setup();
+    const dialog = screen.getByRole("dialog");
+    const closeButtons = within(dialog).getAllByRole("button", {
+      name: "Close",
+    });
+    const textCloseButton = closeButtons.find((btn) => {
+      return btn.textContent?.trim() === "Close";
+    });
+    if (!textCloseButton) {
+      throw new Error("Close button not found");
+    }
+    await user.click(textCloseButton);
+
+    await waitFor(() => {
+      expect(screen.queryByRole("dialog")).not.toBeInTheDocument();
+    });
+  });
+});
